+++ docs/html/search/all_d.js
@@ -1,12 +1,5 @@
 var searchData=
 [
-  ['setpassorfail_153',['setPassOrFail',['../classaunit_1_1Test.html#aec33bdf4a8fd58669d36b233b5364929',1,'aunit::Test']]],
-  ['setprinter_154',['setPrinter',['../classaunit_1_1Printer.html#a2d170283a810ddfec4a0f5744aab56dc',1,'aunit::Printer::setPrinter()'],['../classaunit_1_1TestRunner.html#ac00d7dc7ae426912edce78dc096066ed',1,'aunit::TestRunner::setPrinter()']]],
-  ['setstatus_155',['setStatus',['../classaunit_1_1Test.html#a536f6579a8275e696c76a889dc8d83d5',1,'aunit::Test']]],
-  ['setstatusnow_156',['setStatusNow',['../classaunit_1_1MetaAssertion.html#abaa51ad2cbda0d5ed0ed8b0b162da002',1,'aunit::MetaAssertion']]],
-  ['settimeout_157',['setTimeout',['../classaunit_1_1TestRunner.html#aa22995eb389cd7c5e6e23ccbc7ccbaf6',1,'aunit::TestRunner']]],
-  ['setup_158',['setup',['../classaunit_1_1Test.html#ad32752b60ecc4cab1149d38c2bb2da7c',1,'aunit::Test']]],
-  ['setverbosity_159',['setVerbosity',['../classaunit_1_1TestRunner.html#a70bcaa3a7c0c2a11266ba2c758f4cfc3',1,'aunit::TestRunner']]],
-  ['skip_160',['skip',['../classaunit_1_1Test.html#a8fcf67b6fdec7a1fa364b1b0b07ff1e5',1,'aunit::Test']]],
-  ['skiptestnow_161',['skipTestNow',['../MetaAssertMacros_8h.html#a90a70dc25d2628e1e9353736824117b5',1,'MetaAssertMacros.h']]]
+  ['resolve_154',['resolve',['../classaunit_1_1Test.html#ab7a08fd1e807b989b78d901f2d68ceb5',1,'aunit::Test']]],
+  ['run_155',['run',['../classaunit_1_1TestRunner.html#a93bd4a358d76e551c2aaf0d32d2dff10',1,'aunit::TestRunner']]]
 ];