+++ docs/html/search/all_e.js
@@ -1,13 +1,12 @@
 var searchData=
 [
-  ['teardown_162',['teardown',['../classaunit_1_1Test.html#a698169aed6abd479bd5daec7d1a283c4',1,'aunit::Test']]],
-  ['test_163',['Test',['../classaunit_1_1Test.html',1,'aunit::Test'],['../classaunit_1_1Test.html#a0550ff015d168b10c3c64540081fb19e',1,'aunit::Test::Test()'],['../TestMacros_8h.html#a152fb8fe682506b1b08ddda59abce668',1,'test():&#160;TestMacros.h']]],
-  ['testagain_164',['TestAgain',['../classaunit_1_1TestAgain.html',1,'aunit::TestAgain'],['../classaunit_1_1TestAgain.html#a157fca3287056b91ad022db312ab24d5',1,'aunit::TestAgain::TestAgain()']]],
-  ['testf_165',['testF',['../TestMacros_8h.html#a4f2798c22904efe9442ce65eb6932a9c',1,'TestMacros.h']]],
-  ['testing_166',['testing',['../TestMacros_8h.html#a1f7b0603fa6951be1f019586eafc6e6b',1,'TestMacros.h']]],
-  ['testingf_167',['testingF',['../TestMacros_8h.html#a648c8cb704b9d942b36d8c4646645c2c',1,'TestMacros.h']]],
-  ['testmacros_2eh_168',['TestMacros.h',['../TestMacros_8h.html',1,'']]],
-  ['testonce_169',['TestOnce',['../classaunit_1_1TestOnce.html',1,'aunit::TestOnce'],['../classaunit_1_1TestOnce.html#aca92b171f709cf401701feb9750d8e64',1,'aunit::TestOnce::TestOnce()']]],
-  ['testrunner_170',['TestRunner',['../classaunit_1_1TestRunner.html',1,'aunit']]],
-  ['timeouttype_171',['TimeoutType',['../classaunit_1_1TestRunner.html#a4ea03044dda8ec83aa21b2afcac6e8cf',1,'aunit::TestRunner']]]
+  ['setpassorfail_156',['setPassOrFail',['../classaunit_1_1Test.html#aec33bdf4a8fd58669d36b233b5364929',1,'aunit::Test']]],
+  ['setprinter_157',['setPrinter',['../classaunit_1_1Printer.html#a2d170283a810ddfec4a0f5744aab56dc',1,'aunit::Printer::setPrinter()'],['../classaunit_1_1TestRunner.html#ac00d7dc7ae426912edce78dc096066ed',1,'aunit::TestRunner::setPrinter()']]],
+  ['setstatus_158',['setStatus',['../classaunit_1_1Test.html#a536f6579a8275e696c76a889dc8d83d5',1,'aunit::Test']]],
+  ['setstatusnow_159',['setStatusNow',['../classaunit_1_1MetaAssertion.html#abaa51ad2cbda0d5ed0ed8b0b162da002',1,'aunit::MetaAssertion']]],
+  ['settimeout_160',['setTimeout',['../classaunit_1_1TestRunner.html#aa22995eb389cd7c5e6e23ccbc7ccbaf6',1,'aunit::TestRunner']]],
+  ['setup_161',['setup',['../classaunit_1_1Test.html#ad32752b60ecc4cab1149d38c2bb2da7c',1,'aunit::Test']]],
+  ['setverbosity_162',['setVerbosity',['../classaunit_1_1TestRunner.html#a70bcaa3a7c0c2a11266ba2c758f4cfc3',1,'aunit::TestRunner']]],
+  ['skip_163',['skip',['../classaunit_1_1Test.html#a8fcf67b6fdec7a1fa364b1b0b07ff1e5',1,'aunit::Test']]],
+  ['skiptestnow_164',['skipTestNow',['../MetaAssertMacros_8h.html#a90a70dc25d2628e1e9353736824117b5',1,'MetaAssertMacros.h']]]
 ];