+++ docs/html/search/classes_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['assertion_173',['Assertion',['../classaunit_1_1Assertion.html',1,'aunit']]]
+  ['assertion_176',['Assertion',['../classaunit_1_1Assertion.html',1,'aunit']]]
 ];