+++ docs/html/search/classes_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['printer_177',['Printer',['../classaunit_1_1Printer.html',1,'aunit']]]
+  ['printer_180',['Printer',['../classaunit_1_1Printer.html',1,'aunit']]]
 ];