+++ docs/html/search/classes_4.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['test_178',['Test',['../classaunit_1_1Test.html',1,'aunit']]],
-  ['testagain_179',['TestAgain',['../classaunit_1_1TestAgain.html',1,'aunit']]],
-  ['testonce_180',['TestOnce',['../classaunit_1_1TestOnce.html',1,'aunit']]],
-  ['testrunner_181',['TestRunner',['../classaunit_1_1TestRunner.html',1,'aunit']]]
+  ['test_181',['Test',['../classaunit_1_1Test.html',1,'aunit']]],
+  ['testagain_182',['TestAgain',['../classaunit_1_1TestAgain.html',1,'aunit']]],
+  ['testonce_183',['TestOnce',['../classaunit_1_1TestOnce.html',1,'aunit']]],
+  ['testrunner_184',['TestRunner',['../classaunit_1_1TestRunner.html',1,'aunit']]]
 ];