+++ docs/html/search/functions_a.js
@@ -1,6 +1,4 @@
 var searchData=
 [
-  ['pass_238',['pass',['../classaunit_1_1Test.html#ae71ded07fd6fc69413f64d3d603e4bd6',1,'aunit::Test']]],
-  ['print_239',['print',['../classaunit_1_1internal_1_1FCString.html#ac4a029df196558927f210314c2e28944',1,'aunit::internal::FCString']]],
-  ['println_240',['println',['../classaunit_1_1internal_1_1FCString.html#ae2e214a9db25ed35ae6f448cf133ed1e',1,'aunit::internal::FCString']]]
+  ['once_243',['once',['../classaunit_1_1TestOnce.html#a064675c6adcce5ddb8562628e573ccc1',1,'aunit::TestOnce']]]
 ];